chore move sql_execution_context to sqllab package
